Added full rotation and converted to f64

+++ src/cube.rs
@@ -8,9 +8,9 @@ pub struct Cube {
 }
 
 impl Cube {
-    pub fn new(x: f32, y: f32, z: f32, color: u32, window: &mut three::Window) -> Cube {
+    pub fn new(x: f64, y: f64, z: f64, color: u32, window: &mut three::Window) -> Cube {
         
-        let geometry = three::Geometry::cuboid(x, y, z);
+        let geometry = three::Geometry::cuboid(x as f32, y as f32, z as f32);
         let material = three::material::Basic {
             color: color,
             .. Default::default()
@@ -25,6 +25,6 @@ impl Cube {
     }
 
     pub fn update(&mut self, position: &Vector3f) {
-        self.mesh.set_position([position.x, position.y, position.z]);
+        self.mesh.set_position([position.x as f32, position.y as f32, position.z as f32]);
     }
 }
